feat: collection view 레이아웃 구성하기 #5
